pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "django>=5.2.12",
    "django-filter>=25.2",
    "djangorestframework>=3.16.1",
    "drf-yasg>=1.21.15",
]
